move lines in the buffer before trying to allocate more memory for them.
if we got a lot of lines in a single write, we might have short
line hanging over the end of the buffer. if that buffer is at the
max len, we mistake this as a long line and close the connection.

now we move the short line to the start of the buffer so it has
space at the end.